requirements.txt
graphviz==0.13.2
Pillow==7.0.0
pydotplus==2.0.2
pyfiglet==0.8.post1
pyparsing==2.4.6
